AutoCheckin

+++ src/mat/order/degree.c
@@ -143,7 +143,7 @@ L100:
 /*<             JSTRT = -XADJ(NODE)                                          >*/
 	jstrt = -xadj[node];
 /*<             JSTOP = IABS(XADJ(NODE + 1)) - 1                             >*/
-	jstop = (i__2 = xadj[node + 1], abs(i__2)) - 1;
+	jstop = (i__2 = xadj[node + 1], (int)abs(i__2)) - 1;
 /*<             IDEG = 0                                                     >*/
 	ideg = 0;
 /*<             IF ( JSTOP .LT. JSTRT ) GO TO 300                            >*/
