+++ src/mat/order/fndsep.c
@@ -191,7 +191,7 @@ L200:
 /*<             JSTRT = XADJ(NODE)                                           >*/
 	jstrt = xadj[node];
 /*<             JSTOP = IABS(XADJ(NODE+1)) - 1                               >*/
-	jstop = (i__2 = xadj[node + 1], abs(i__2)) - 1;
+	jstop = (i__2 = xadj[node + 1], (int)abs(i__2)) - 1;
 /*<             DO 400 J = JSTRT, JSTOP                                      >*/
 	i__2 = jstop;
 	for (j = jstrt; j <= i__2; ++j) {
